requirements.txt
Flask==0.12.2
flask-mongoengine==0.9.3
nose==1.3.7
numpy==1.13
scipy==0.19.1
requests==2.18.4